backend/requirements.txt
google-genai>=1.0.0
ag2[gemini]>=0.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fal-client>=0.5.0
elevenlabs>=1.0.0
httpx>=0.27.0
